+++ my_app/reward-calculation-app/src/main/java/org/javaguru/reward/calculation/rest/employee/EmployeeDTO.java
@@ -9,6 +9,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -21,5 +23,5 @@ public class EmployeeDTO {
 
     private String lastName;
 
-    private Double bonusCoefficient;
+    private BigDecimal bonusCoefficient;
 }
