+++ my_app/reward-calculation-app/src/main/java/org/javaguru/reward/calculation/service/restclient/RewardPaymentClient.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
 
+import java.math.BigDecimal;
+
 @Service
 @RequiredArgsConstructor(access = AccessLevel.PACKAGE)
 public class RewardPaymentClient {
@@ -13,7 +15,7 @@ public class RewardPaymentClient {
 //    Создается в RewardPaymentClientConfig
     private final RestClient rewardPaymentRestClient;
 
-    public void payReward(Long employeeId, Double amount) {
+    public void payReward(Long employeeId, BigDecimal amount) {
         RewardPaymentRequest request = new RewardPaymentRequest(employeeId, amount);
 
         rewardPaymentRestClient.post()
